refactor: rename 'server' folder to 'backend'
